requirements.txt
Flask==3.0.0
praw==7.8.1
python-dotenv==1.0.0
gunicorn==21.2.0